+++ R/checkers.R
@@ -59,9 +59,7 @@ check_any_missing_countries <- function(
     cli::cli_abort(
       message = c(
         "Could not find population data for: {.val {missing_countries}}.",
-        # nolint start
-        "i" = "Use {.fn wpp_countries} to get a list of country names."
-        # nolint end
+        i = "Pass population data directly via the {.arg survey_pop} argument."
       ),
       call = call
     )